+++ src/Testcontainers/Volumes/DockerVolume.cs
@@ -1,5 +1,6 @@
 namespace DotNet.Testcontainers.Volumes
 {
+  using System;
   using System.Threading;
   using System.Threading.Tasks;
   using Docker.DotNet.Models;
@@ -42,6 +43,57 @@ namespace DotNet.Testcontainers.Volumes
     /// <inheritdoc />
     public async Task CreateAsync(CancellationToken ct = default)
     {
+      using (_ = this.AcquireLock())
+      {
+        await this.UnsafeCreateAsync(ct)
+          .ConfigureAwait(false);
+      }
+    }
+
+    /// <inheritdoc />
+    public async Task DeleteAsync(CancellationToken ct = default)
+    {
+      using (_ = this.AcquireLock())
+      {
+        await this.UnsafeDeleteAsync(ct)
+          .ConfigureAwait(false);
+      }
+    }
+
+    /// <inheritdoc />
+    protected override async ValueTask DisposeAsyncCore()
+    {
+      if (this.Disposed)
+      {
+        return;
+      }
+
+      if (!Guid.Empty.Equals(this.configuration.SessionId))
+      {
+        await this.DeleteAsync()
+          .ConfigureAwait(false);
+      }
+
+      await base.DisposeAsyncCore()
+        .ConfigureAwait(false);
+    }
+
+    /// <inheritdoc />
+    protected override bool Exists()
+    {
+      return !string.IsNullOrEmpty(this.volume.Name);
+    }
+
+    /// <inheritdoc />
+    protected override async Task UnsafeCreateAsync(CancellationToken ct = default)
+    {
+      this.ThrowIfLockNotAcquired();
+
+      if (this.Exists())
+      {
+        return;
+      }
+
       var name = await this.dockerVolumeOperations.CreateAsync(this.configuration, ct)
         .ConfigureAwait(false);
 
@@ -50,18 +102,19 @@ namespace DotNet.Testcontainers.Volumes
     }
 
     /// <inheritdoc />
-    public async Task DeleteAsync(CancellationToken ct = default)
+    protected override async Task UnsafeDeleteAsync(CancellationToken ct = default)
     {
+      this.ThrowIfLockNotAcquired();
+
+      if (!this.Exists())
+      {
+        return;
+      }
+
       await this.dockerVolumeOperations.DeleteAsync(this.Name, ct)
         .ConfigureAwait(false);
 
       this.volume = new VolumeResponse();
-    }
-
-    /// <inheritdoc />
-    protected override bool Exists()
-    {
-      return !string.IsNullOrEmpty(this.volume.Name);
     }
   }
 }
